feat(app): create a minimum bevy app.
